Remove old bftcosi
